+++ include/triton/Conversion/TritonGPUToLLVM/Passes.h
@@ -1,16 +1,36 @@
-#ifndef TRITONGPU_CONVERSION_PASSES_H
-#define TRITONGPU_CONVERSION_PASSES_H
+#ifndef TRITONGPU_CONVERSION_TRITONGPUTOLLVM_PASSES_H
+#define TRITONGPU_CONVERSION_TRITONGPUTOLLVM_PASSES_H
 
-#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
-#include "triton/Conversion/TritonGPUToLLVM/TritonGPUToLLVMPass.h"
+#include "mlir/Conversion/LLVMCommon/TypeConverter.h"
+#include "mlir/Transforms/DialectConversion.h"
+#include "triton/Target/PTX/TmaMetadata.h"
+
+#include <memory>
 
 namespace mlir {
+
+class ModuleOp;
+template <typename T> class OperationPass;
+
 namespace triton {
 
-#define GEN_PASS_REGISTRATION
+enum Target { NVVM, ROCDL, Default = NVVM };
+
+#define GEN_PASS_DECL
 #include "triton/Conversion/TritonGPUToLLVM/Passes.h.inc"
 
+namespace gpu {
+std::unique_ptr<OperationPass<ModuleOp>>
+createDecomposeUnsupportedConversionsPass();
+} // namespace gpu
+
+std::unique_ptr<OperationPass<ModuleOp>> createConvertTritonGPUToLLVMPass();
+std::unique_ptr<OperationPass<ModuleOp>>
+createConvertTritonGPUToLLVMPass(int32_t computeCapability, Target target,
+                                 mlir::triton::gpu::TMAMetadataTy *tmaMetadata);
+
 } // namespace triton
+
 } // namespace mlir
 
 #endif
