+++ include/triton/Conversion/TritonGPUToLLVM/Passes.td
@@ -3,7 +3,6 @@
 
 include "mlir/Pass/PassBase.td"
 
-
 def DecomposeUnsupportedConversions : Pass<"decompose-unsupported-conversions", "mlir::ModuleOp"> {
     let summary = "Decompose conversions that are not supported by TritonGPU -> LLVM";
     let constructor = "mlir::triton::gpu::createDecomposeUnsupportedConversionsPass()";
