+++ lib/Conversion/TritonGPUToLLVM/CMakeLists.txt
@@ -14,6 +14,7 @@ add_triton_library(TritonGPUToLLVM
     BarrierOpToLLVM.cpp
     TritonGPUToLLVM.cpp
     TritonGPUToLLVMPass.cpp
+    DecomposeUnsupportedConversions.cpp
     ReduceOpToLLVM.cpp
     ScanOpToLLVM.cpp
     TypeConverter.cpp
